Calculator.java and CalculatorMain.java updated

+++ training-solutions/src/main/java/introexceptiontrycatchtrace/Calculator.java
@@ -2,19 +2,19 @@ package introexceptiontrycatchtrace;
 
 public class Calculator {
 
-    public void add(int number1,int number2){
+    public void add(double number1,double number2){
         System.out.println(number1+number2);
     }
 
-    public void subtract(int number1,int number2){
+    public void subtract(double number1,double number2){
         System.out.println(number1-number2);
     }
 
-    public void multiply(int number1,int number2){
+    public void multiply(double number1,double number2){
         System.out.println(number1*number2);
     }
 
     public void divide(int number1,int number2){
-        System.out.println((double)(number1/number2));
+        System.out.println(number1/number2);
     }
 }
